PaymentTypes Migration Unique and Transfer Login into LoginController
