+++ backend/app/Http/Controllers/UserController.php
@@ -10,8 +10,6 @@ use App\Http\Resources\UserResource;
 use Illuminate\Support\Facades\Hash;
 use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
-use App\Http\Requests\LoginRequest;
-use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -110,35 +108,5 @@ class UserController extends Controller
         ]);
     }
     
-    public function login(LoginRequest $request)
-    {
-        $credentials = $request->only('email', 'password');
 
-        if (!Auth::attempt($credentials)) {
-            return response()->json([
-                'message' => 'The email and password does not match in our database!',
-                'success' => false
-            ], 401);
-        }
-
-        $user = User::where('email', $request->email)->first();
-        $token = $user->createToken('user-token');
-
-        return response()->json([
-            'data' => new UserResource($user->load(['userDetail', 'roles'])),
-            'access_token' => $token->plainTextToken,
-            'token_type' => 'Bearer'
-        ]);
-    }
-
-    public function logout()
-    {
-        $user = Auth::user();
-        $user->tokens()->delete();
-
-        return response()->json([
-            'message' => 'The user was succesfully logged out!',
-            'success' => true
-        ]);
-    }
 }
